[FEAT] UserService 추상화 및 findByEncryptedId 구현
