+++ src/main/java/synk/meeteam/domain/user/user/service/UserService.java
@@ -1,65 +1,16 @@
 package synk.meeteam.domain.user.user.service;
 
-import static synk.meeteam.domain.user.user.exception.UserExceptionType.DUPLICATE_NICKNAME;
-
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
-import synk.meeteam.domain.common.role.entity.Role;
-import synk.meeteam.domain.common.role.repository.RoleRepository;
 import synk.meeteam.domain.user.user.dto.command.UpdateInfoCommand;
 import synk.meeteam.domain.user.user.entity.User;
-import synk.meeteam.domain.user.user.exception.UserException;
-import synk.meeteam.domain.user.user.repository.UserRepository;
 
-@Service
-@RequiredArgsConstructor
-public class UserService {
+public interface UserService {
+    boolean checkAvailableNickname(String nickname);
 
-    private final UserRepository userRepository;
-    private final RoleRepository roleRepository;
+    void changeOrKeepNickname(User user, String newNickname);
 
-    @Transactional(readOnly = true)
-    public boolean checkAvailableNickname(String nickname) {
-        User foundUser = userRepository.findByNickname(nickname).orElse(null);
-        return foundUser == null;
-    }
+    void changeUserInfo(User user, UpdateInfoCommand command);
 
-    @Transactional
-    public void changeOrKeepNickname(User user, String newNickname) {
-        if (user.isNotEqualNickname(newNickname)) {
-            if (checkAvailableNickname(newNickname)) {
-                user.updateNickname(newNickname);
-            } else {
-                throw new UserException(DUPLICATE_NICKNAME);
-            }
-        }
-    }
+    User findById(Long userId);
 
-    @Transactional
-    public void changeUserInfo(User user, UpdateInfoCommand command) {
-
-        Role interestRole = roleRepository.findByIdOrElseThrowException(command.interest_id());
-        user.updateProfile(
-                command.name(),
-                command.pictureUrl(),
-                command.subEmail(),
-                command.isPublicSubEmail(),
-                command.isPublicSchoolEmail(),
-                command.isSchoolMain(),
-                command.phoneNumber(),
-                command.isPublicPhone(),
-                command.oneLineIntroduction(),
-                command.mainIntroduction(),
-                command.gpa(),
-                command.maxGpa(),
-                interestRole
-        );
-        userRepository.save(user);
-    }
-
-    @Transactional(readOnly = true)
-    public User findById(Long userId) {
-        return userRepository.findByIdOrElseThrow(userId);
-    }
+    User findByEncryptedId(String encryptedId);
 }
